Move tests from vtag to classes
