+++ yew/src/virtual_dom/classes.rs
@@ -225,3 +225,402 @@ mod tests {
         assert!(subject.contains("bar"));
     }
 }
+
+#[cfg(test)]
+mod macro_tests {
+    extern crate self as yew;
+
+    use crate::virtual_dom::{AnyScope, VDiff, VNode, VTag};
+    use crate::{classes, html, NodeRef};
+    use std::any::TypeId;
+    use std::rc::Rc;
+    #[cfg(feature = "std_web")]
+    use stdweb::web::{document, IElement};
+    #[cfg(feature = "wasm_test")]
+    use wasm_bindgen_test::{wasm_bindgen_test as test, wasm_bindgen_test_configure};
+
+    #[cfg(feature = "wasm_test")]
+    wasm_bindgen_test_configure!(run_in_browser);
+
+    fn test_scope() -> AnyScope {
+        AnyScope {
+            type_id: TypeId::of::<()>(),
+            parent: None,
+            state: Rc::new(()),
+        }
+    }
+
+    #[test]
+    fn classes_from_local_variables() {
+        let a = html! {
+            <div class=classes!("class-1", "class-2")></div>
+        };
+
+        let class_2 = "class-2";
+        let b = html! {
+            <div class=classes!("class-1", class_2)></div>
+        };
+
+        let class_2_fmt = format!("class-{}", 2);
+        let c = html! {
+            <div class=classes!("class-1", class_2_fmt)></div>
+        };
+
+        assert_eq!(a, b);
+        assert_eq!(a, c);
+    }
+
+    /// Returns the class attribute as str reference, or "" if the attribute is not set.
+    fn get_class_str(vtag: &VTag) -> &str {
+        vtag.attributes
+            .iter()
+            .find(|(k, _)| k == &"class")
+            .map(|(_, v)| AsRef::as_ref(v))
+            .unwrap_or("")
+    }
+
+    /// Note: Compares to "" if the class attribute is not set.
+    fn assert_class(vnode: VNode, class: &str) {
+        if let VNode::VTag(ref vtag) = vnode {
+            assert_eq!(get_class_str(vtag), class);
+        } else {
+            panic!("expected VTag");
+        }
+    }
+
+    #[test]
+    fn supports_multiple_non_unique_classes_tuple() {
+        let a = html! {
+            <div class=classes!("class-1", "class-1 class-2")></div>
+        };
+
+        if let VNode::VTag(vtag) = a {
+            assert!(get_class_str(&vtag).contains("class-1"));
+            assert!(get_class_str(&vtag).contains("class-2"));
+            assert!(!get_class_str(&vtag).contains("class-3"));
+        } else {
+            panic!("vtag expected");
+        }
+    }
+
+    #[test]
+    fn supports_multiple_classes_string() {
+        let a = html! {
+            <div class=classes!("class-1 class-2 class-3")></div>
+        };
+        let b = html! {
+            <div class=classes!("class-2 class-3 class-1")></div>
+        };
+
+        assert_ne!(a, b);
+
+        if let VNode::VTag(vtag) = a {
+            assert!(get_class_str(&vtag).contains("class-1"));
+            assert!(get_class_str(&vtag).contains("class-2"));
+            assert!(get_class_str(&vtag).contains("class-3"));
+        } else {
+            panic!("vtag expected");
+        }
+    }
+
+    #[test]
+    fn supports_multiple_classes_slice() {
+        let classes = classes!(&["class-1", "class-2"][..]);
+        let a = html! {
+            <div class=classes></div>
+        };
+
+        if let VNode::VTag(vtag) = a {
+            assert!(get_class_str(&vtag).contains("class-1"));
+            assert!(get_class_str(&vtag).contains("class-2"));
+            assert!(!get_class_str(&vtag).contains("class-3"));
+        } else {
+            panic!("vtag expected");
+        }
+    }
+
+    #[test]
+    fn supports_multiple_classes_one_value_slice() {
+        let classes = classes!(&["class-1 class-2", "class-1"][..]);
+        let a = html! {
+            <div class=classes></div>
+        };
+
+        if let VNode::VTag(vtag) = a {
+            assert!(get_class_str(&vtag).contains("class-1"));
+            assert!(get_class_str(&vtag).contains("class-2"));
+            assert!(!get_class_str(&vtag).contains("class-3"));
+        } else {
+            panic!("vtag expected");
+        }
+    }
+
+    #[test]
+    fn supports_multiple_classes_vec() {
+        let mut classes = vec!["class-1"];
+        classes.push("class-2");
+        let classes = classes!(classes);
+        let a = html! {
+            <div class=classes></div>
+        };
+
+        if let VNode::VTag(vtag) = a {
+            assert!(get_class_str(&vtag).contains("class-1"));
+            assert!(get_class_str(&vtag).contains("class-2"));
+            assert!(!get_class_str(&vtag).contains("class-3"));
+        } else {
+            panic!("vtag expected");
+        }
+    }
+
+    #[test]
+    fn supports_multiple_classes_one_value_vec() {
+        let classes = vec!["class-1 class-2", "class-1"];
+        let classes = classes!(classes);
+        let a = html! {
+            <div class=classes></div>
+        };
+
+        if let VNode::VTag(vtag) = a {
+            assert!(get_class_str(&vtag).contains("class-1"));
+            assert!(get_class_str(&vtag).contains("class-2"));
+            assert!(!get_class_str(&vtag).contains("class-3"));
+        } else {
+            panic!("vtag expected");
+        }
+    }
+
+    #[test]
+    fn filter_empty_string_classes() {
+        let a = html! { <div class=classes!(vec![""])></div> };
+        let b = html! { <div class=classes!("", "")></div> };
+        let c = html! { <div class=""></div> };
+        let d_arr = [""];
+        let d = html! { <div class=classes!(&d_arr[..])></div> };
+
+        macro_rules! get_class {
+            ($vtag:expr) => {
+                $vtag
+                    .attributes
+                    .iter()
+                    .find_map(|(k, v)| if k == "class" { Some(v) } else { None })
+            };
+        }
+
+        if let VNode::VTag(vtag) = a {
+            assert_eq!(get_class!(vtag), None);
+        } else {
+            panic!("vtag expected");
+        }
+
+        if let VNode::VTag(vtag) = b {
+            assert_eq!(get_class!(vtag), None);
+        } else {
+            panic!("vtag expected");
+        }
+
+        if let VNode::VTag(vtag) = c {
+            assert_eq!(get_class!(vtag), None);
+        } else {
+            panic!("vtag expected");
+        }
+
+        if let VNode::VTag(vtag) = d {
+            assert_eq!(get_class!(vtag), None);
+        } else {
+            panic!("vtag expected");
+        }
+    }
+
+    #[test]
+    fn keeps_order_of_classes() {
+        let a = html! {
+            <div class=classes!(vec!["class-1", "class-2", "class-3"])></div>
+        };
+
+        if let VNode::VTag(vtag) = a {
+            assert_eq!(get_class_str(&vtag), "class-1 class-2 class-3");
+        }
+    }
+
+    #[test]
+    fn tuple_different_types() {
+        // check if tuples containing different types are compiling
+        assert_class(
+            html! { <div class=classes!("class-1", "class-2".to_string(), vec!["class-3", "class-4"])></div> },
+            "class-1 class-2 class-3 class-4",
+        );
+        assert_class(
+            html! { <div class=classes!("class-1", Some("class-2"), "class-3", Some("class-4".to_string()))></div> },
+            "class-1 class-2 class-3 class-4",
+        );
+        // check different string references
+        let str = "some-class";
+        let string = str.to_string();
+        let string_ref = &string;
+        assert_class(html! { <p class=classes!(str) /> }, "some-class");
+        assert_class(html! { <p class=classes!(string.clone()) /> }, "some-class");
+        assert_class(html! { <p class=classes!(&Some(str)) /> }, "some-class");
+        assert_class(html! { <p class=classes!(string_ref) /> }, "some-class");
+        assert_class(html! { <p class=classes!(Some(str)) /> }, "some-class");
+        assert_class(
+            html! { <p class=classes!(Some(string.clone())) /> },
+            "some-class",
+        );
+        assert_class(
+            html! { <p class=classes!(Some(string_ref)) /> },
+            "some-class",
+        );
+        assert_class(
+            html! { <p class=classes!(&Some(string.clone())) /> },
+            "some-class",
+        );
+        assert_class(
+            html! { <p class=classes!(&Some(string_ref)) /> },
+            "some-class",
+        );
+        // check with None
+        assert_class(html! { <p class=classes!(&Option::<&str>::None) /> }, "");
+        assert_class(html! { <p class=classes!(Option::<String>::None) /> }, "");
+        // check with variables
+        let some: Option<&'static str> = Some("some");
+        let none: Option<&'static str> = None;
+        assert_class(html! { <p class=classes!(some) /> }, "some");
+        assert_class(html! { <p class=classes!(none) /> }, "");
+        // check with variables of different type
+        let some: Option<bool> = Some(false);
+        let none: Option<bool> = None;
+        assert_class(
+            html! { <p class=classes!(some.map(|i| i.to_string())) /> },
+            "false",
+        );
+        assert_class(
+            html! { <p class=classes!(none.map(|i| i.to_string())) /> },
+            "",
+        );
+    }
+
+    #[test]
+    #[cfg(any(feature = "std_web", feature = "web_sys"))]
+    fn swap_order_of_classes() {
+        #[cfg(feature = "std_web")]
+        let document = document();
+        #[cfg(feature = "web_sys")]
+        let document = web_sys::window().unwrap().document().unwrap();
+
+        let scope = test_scope();
+        let parent = document.create_element("div").unwrap();
+
+        #[cfg(feature = "std_web")]
+        document.body().unwrap().append_child(&parent);
+        #[cfg(feature = "web_sys")]
+        document.body().unwrap().append_child(&parent).unwrap();
+
+        let mut elem = html! { <div class=classes!("class-1", "class-2", "class-3")></div> };
+        elem.apply(&scope, &parent, NodeRef::default(), None);
+
+        let vtag = if let VNode::VTag(vtag) = elem {
+            vtag
+        } else {
+            panic!("should be vtag")
+        };
+
+        let expected = "class-1 class-2 class-3";
+        assert_eq!(get_class_str(&vtag), expected);
+        assert_eq!(
+            vtag.reference
+                .as_ref()
+                .unwrap()
+                .get_attribute("class")
+                .unwrap(),
+            expected
+        );
+
+        let ancestor = vtag;
+        let elem = html! { <div class=classes!("class-3", "class-2", "class-1")></div> };
+        let mut vtag = if let VNode::VTag(vtag) = elem {
+            vtag
+        } else {
+            panic!("should be vtag")
+        };
+        vtag.apply(
+            &scope,
+            &parent,
+            NodeRef::default(),
+            Some(VNode::VTag(ancestor)),
+        );
+
+        let expected = "class-3 class-2 class-1";
+        assert_eq!(get_class_str(&vtag), expected);
+        assert_eq!(
+            vtag.reference
+                .as_ref()
+                .unwrap()
+                .get_attribute("class")
+                .unwrap(),
+            expected
+        );
+    }
+
+    #[test]
+    #[cfg(any(feature = "std_web", feature = "web_sys"))]
+    fn add_class_to_the_middle() {
+        #[cfg(feature = "std_web")]
+        let document = document();
+        #[cfg(feature = "web_sys")]
+        let document = web_sys::window().unwrap().document().unwrap();
+
+        let scope = test_scope();
+        let parent = document.create_element("div").unwrap();
+
+        #[cfg(feature = "std_web")]
+        document.body().unwrap().append_child(&parent);
+        #[cfg(feature = "web_sys")]
+        document.body().unwrap().append_child(&parent).unwrap();
+
+        let mut elem = html! { <div class=classes!("class-1", "class-3")></div> };
+        elem.apply(&scope, &parent, NodeRef::default(), None);
+
+        let vtag = if let VNode::VTag(vtag) = elem {
+            vtag
+        } else {
+            panic!("should be vtag")
+        };
+
+        let expected = "class-1 class-3";
+        assert_eq!(get_class_str(&vtag), expected);
+        assert_eq!(
+            vtag.reference
+                .as_ref()
+                .unwrap()
+                .get_attribute("class")
+                .unwrap(),
+            expected
+        );
+
+        let ancestor = vtag;
+        let elem = html! { <div class=classes!("class-1", "class-2", "class-3")></div> };
+        let mut vtag = if let VNode::VTag(vtag) = elem {
+            vtag
+        } else {
+            panic!("should be vtag")
+        };
+        vtag.apply(
+            &scope,
+            &parent,
+            NodeRef::default(),
+            Some(VNode::VTag(ancestor)),
+        );
+
+        let expected = "class-1 class-2 class-3";
+        assert_eq!(get_class_str(&vtag), expected);
+        assert_eq!(
+            vtag.reference
+                .as_ref()
+                .unwrap()
+                .get_attribute("class")
+                .unwrap(),
+            expected
+        );
+    }
+}
